[fstd] Adding some performance tests to check if to_string is fast enough (which it isn't actually)

+++ compiler-cpp/sources/fstd/system/file.cpp
@@ -79,7 +79,7 @@ namespace fstd
 			if (ReadFile(
 				file.handle,						// hFile
 				memory::get_array_data(content),	// lpBuffer
-				memory::get_array_size(content),	// nNumberOfBytesToRead
+				(DWORD)memory::get_array_size(content),	// nNumberOfBytesToRead
 				&read,								// lpNumberOfBytesRead
 				NULL								// lpOverlapped
 			) == TRUE) {
